Remove OptNode class
