+++ tests/golem/docker/test_docker_dummy_task.py
@@ -12,7 +12,7 @@ from golem.clientconfigdescriptor import ClientConfigDescriptor
 from golem.core.common import get_golem_path, timeout_to_deadline
 from golem.core.fileshelper import find_file_with_ext
 from golem.core.simpleserializer import DictSerializer
-from golem.node import OptNode
+from golem.node import Node
 from golem.resource.dirmanager import DirManager, symlink_or_copy, \
     rmlink_or_rmtree
 from golem.task.localcomputer import LocalComputer
@@ -124,7 +124,7 @@ class TestDockerDummyTask(TempDirFixture, DockerTestCase):
         ctd['deadline'] = timeout_to_deadline(timeout)
 
         # Create the computing node
-        self.node = OptNode(datadir=self.path, use_docker_machine_manager=False)
+        self.node = Node(datadir=self.path, use_docker_machine_manager=False)
         self.node.client.start = Mock()
         self.node.client.datadir = self.path
         self.node._run()
